requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argcomplete==2.0.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Flask==2.1.0
frozenlist==1.3.0
git+https://github.com/uniswap-python/uniswap-python@master
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.11.3
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pipx==1.0.0
protobuf==3.19.4
pycryptodome==3.14.1
pyee==9.0.4
pyparsing==3.0.6
pyrsistent==0.18.1
python-dotenv==0.19.2
requests==2.27.1
rlp==2.0.1
six==1.16.0
slack==0.0.2
slackclient==2.9.3
slackeventsapi==3.0.1
toolz==0.11.2
typing-extensions==4.1.1
urllib3==1.26.8
userpath==1.8.0
varint==1.0.2
web3==5.28.0
websockets==9.1
Werkzeug==2.1.0
yarl==1.7.2
zipp==3.7.0
hvac==0.10.5
